+++ template-presets/localization/index.ts
@@ -1,0 +1,2 @@
+export { useLocalization } from "./provider";
+export { LocalizationProvider } from "./provider";